skip STA if no CLOCK_PORT is defined

+++ scripts/tcl_commands/synthesis.tcl
@@ -59,8 +59,12 @@ proc run_yosys {args} {
 
 proc run_sta {args} {
     puts_info "Running Static Timing Analysis..."
-    try_catch sta $::env(SCRIPTS_DIR)/sta.tcl \
-	|& tee $::env(TERMINAL_OUTPUT) $::env(opensta_log_file_tag).log
+    if {[info exists ::env(CLOCK_PORT)]} {
+        try_catch sta $::env(SCRIPTS_DIR)/sta.tcl \
+        |& tee $::env(TERMINAL_OUTPUT) $::env(opensta_log_file_tag).log
+    } else {
+        puts_warn "No CLOCK_PORT found. Skipping STA..."
+    }
 }
 
 proc run_synth_exploration {args} {
